♻️ ProblemDto 클래스를 추상 클래스로 변경하고, 문제 유형에 따라 상속된 DTO로 변환하도록 수정
