+++ src/main/java/site/haruhana/www/dto/problem/extend/SubjectiveProblemDto.java
@@ -1,0 +1,17 @@
+package site.haruhana.www.dto.problem.extend;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Getter;
+import site.haruhana.www.dto.problem.ProblemDto;
+import site.haruhana.www.entity.problem.Problem;
+
+@Getter
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class SubjectiveProblemDto extends ProblemDto {
+    private final String expectedAnswerLength;
+
+    public SubjectiveProblemDto(Problem problem, Boolean isSolved) {
+        super(problem, isSolved);
+        this.expectedAnswerLength = problem.getExpectedAnswerLength();
+    }
+}
